+++ src/features/config/appConfig.test.ts
@@ -11,6 +11,12 @@ describe('appConfig helpers', () => {
         middleClickOpenTarget: 'blank',
         defaultSearchEngine: 'google',
         customSearchEngines: [],
+        networkProbe: {
+          lanProtocol: 'http',
+          lanHost: '',
+          wanProtocol: 'https',
+          wanHost: '',
+        },
         webdavBackup: {
           url: '',
           username: '',